# openapi.yaml
openapi: 3.1.0
info:
  title: PACEXT — API UniCatólica
  description: >
    Contrato REST da rede social acadêmica UniCatólica (PACEXT). Fonte de verdade
    acordada entre frontend e backend antes de qualquer lado implementar um endpoint
    novo (AD-4 da Architecture Spine) — nenhum lado avança sem que o endpoint exista
    aqui primeiro.
  version: '0.1.0'

servers:
  - url: http://localhost:8080
    description: Ambiente local (docker-compose, AD-7)

paths:
  /auth/login:
    post:
      operationId: login
      summary: Autentica um usuário e emite uma sessão JWT (RF06/RF07/RF08).
      tags: [Identidade]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Credenciais válidas — sessão emitida.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/NaoAutenticado'

  /auth/registro:
    post:
      summary: Cadastro de aluno com e-mail institucional (Story 1.2)
      description: >
        Rota pública (allowlist AD-2). Persiste o usuário com status "e-mail não
        confirmado" — nenhuma sessão é gerada — e dispara o auto-join à comunidade de
        curso correspondente (RF24.1).
      operationId: cadastrarAluno
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CadastroRequest'
      responses:
        '201':
          description: Usuário cadastrado, e-mail de confirmação enviado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CadastroResponse'
        # 409: e-mail já possui conta (RF02). 422: domínio externo ao institucional
        # (RF01.3), formato de e-mail/senha fora da política (RF03/RF04), ou idade menor
        # que 18 anos (RF01.1) — o schema Erro.code compartilhado (ver ErroResponse no
        # backend) distingue os cenários; o response object AD-5 é sempre o mesmo por status.
        '409':
          $ref: '#/components/responses/Conflito'
        '422':
          $ref: '#/components/responses/ErroValidacao'

  /auth/confirmacao-email/{token}:
    post:
      summary: Confirma o e-mail a partir do link recebido (Story 1.3)
      description: Rota pública (allowlist AD-2). Idempotente para token já confirmado.
      operationId: confirmarEmail
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: E-mail confirmado — usuário já pode fazer login.
        # 404: token inexistente ou já substituído por um reenvio (RF01.2). 422: token expirado.
        '404':
          $ref: '#/components/responses/NaoEncontrado'
        '422':
          $ref: '#/components/responses/ErroValidacao'

  /auth/confirmacao-email/reenvio:
    post:
      summary: Reenvia o e-mail de confirmação (Story 1.3)
      description: >
        Rota pública (allowlist AD-2). Sempre responde 202, mesmo para e-mail
        inexistente ou já confirmado — nunca revela se a conta existe.
      operationId: reenviarConfirmacaoEmail
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReenvioConfirmacaoRequest'
      responses:
        '202':
          description: Solicitação aceita.

  /usuarios/me:
    get:
      operationId: obterUsuarioAutenticado
      summary: Retorna os dados do usuário autenticado (RF12).
      tags: [Identidade]
      responses:
        '200':
          description: Dados do próprio usuário.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        '401':
          $ref: '#/components/responses/NaoAutenticado'
        '404':
          $ref: '#/components/responses/NaoEncontrado'

  /usuarios/{id}:
    get:
      operationId: obterUsuarioPorId
      summary: Retorna os dados de um usuário por id — somente perfil MODERADOR (RF13).
      tags: [Identidade]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Dados do usuário solicitado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        '401':
          $ref: '#/components/responses/NaoAutenticado'
        '403':
          $ref: '#/components/responses/SemPermissao'
        '404':
          $ref: '#/components/responses/NaoEncontrado'

  /comunidades:
    post:
      operationId: criarComunidade
      summary: Cria uma comunidade aberta (Story 2.2, RF21/RF22/RF23)
      tags: [Comunidades]
      description: >
        Toda comunidade criada por aqui é do tipo ABERTA — o corpo não aceita "tipo"
        de propósito (RF21.2: só administrador de plataforma cria comunidade de curso,
        endpoint fora desta fatia, ver docs/modelo-epico-2-comunidades.md). O criador
        vira administrador da comunidade automaticamente (RF23).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComunidadeRequest'
      responses:
        '201':
          description: Comunidade criada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComunidadeResponse'
        '401':
          $ref: '#/components/responses/NaoAutenticado'
        # 409: já existe comunidade com esse nome. 422: nome não informado (RF22).
        '409':
          $ref: '#/components/responses/Conflito'
        '422':
          $ref: '#/components/responses/ErroValidacao'
    get:
      operationId: listarComunidades
      summary: Lista e filtra comunidades, paginado (Story 2.5, RF27/RF28)
      tags: [Comunidades]
      parameters:
        - name: tipo
          in: query
          schema:
            type: string
            enum: [CURSO, ABERTA]
        - name: nome
          in: query
          schema:
            type: string
        - name: pagina
          in: query
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
        - name: tamanho
          in: query
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Página de comunidades.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComunidadePageResponse'
        '401':
          $ref: '#/components/responses/NaoAutenticado'

  /comunidades/minhas:
    get:
      operationId: listarMinhasComunidades
      summary: Lista as comunidades das quais o usuário autenticado é membro (Home)
      tags: [Comunidades]
      description: >
        Sem paginação de propósito — alimenta a seção "Suas comunidades" da Home,
        tende a ser um conjunto pequeno.
      responses:
        '200':
          description: Comunidades do usuário autenticado.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ComunidadeResponse'
        '401':
          $ref: '#/components/responses/NaoAutenticado'

  /comunidades/{id}:
    get:
      operationId: obterComunidade
      summary: Detalhe de uma comunidade, incluindo se o usuário é membro (Story 2.5, RF27.1)
      tags: [Comunidades]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Dados da comunidade.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComunidadeResponse'
        '401':
          $ref: '#/components/responses/NaoAutenticado'
        '404':
          $ref: '#/components/responses/NaoEncontrado'

  /comunidades/{id}/membros:
    post:
      operationId: ingressarNaComunidade
      summary: Entra numa comunidade aberta (Story 2.4, RF24/RF25)
      tags: [Comunidades]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '201':
          description: Ingresso realizado.
        '401':
          $ref: '#/components/responses/NaoAutenticado'
        # 403: comunidade não é do tipo ABERTA. 409: já é membro (RF25).
        '403':
          $ref: '#/components/responses/SemPermissao'
        '404':
          $ref: '#/components/responses/NaoEncontrado'
        '409':
          $ref: '#/components/responses/Conflito'

  /comunidades/{id}/membros/me:
    delete:
      operationId: sairDaComunidade
      summary: Sai de uma comunidade (Story 2.4, RF26)
      tags: [Comunidades]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Associação removida.
        '401':
          $ref: '#/components/responses/NaoAutenticado'
        '404':
          $ref: '#/components/responses/NaoEncontrado'

components:
  schemas:
    LoginRequest:
      type: object
      required: [email, senha]
      properties:
        email:
          type: string
          format: email
        senha:
          type: string
          format: password

    LoginResponse:
      type: object
      description: >
        Token JWT com claims sub (id do usuário) e roles (perfil global) — transportado
        daqui em diante via header Authorization: Bearer, nunca cookie (AD-2).
      required: [token]
      properties:
        token:
          type: string

    CadastroRequest:
      type: object
      required:
        - nome
        - email
        - senha
        - curso
        - dataNascimento
      properties:
        nome:
          type: string
        email:
          type: string
          format: email
          description: E-mail institucional (RF01.3) — domínio validado no backend.
        senha:
          type: string
          format: password
        curso:
          type: string
          description: >
            Nome do curso — usado para o auto-join (RF24.1) contra a comunidade de curso
            pré-criada pelo administrador (Story 2.1).
        dataNascimento:
          type: string
          format: date

    CadastroResponse:
      type: object
      required:
        - id
        - nome
        - email
        - curso
        - emailConfirmado
        - criadoEm
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        email:
          type: string
          format: email
        curso:
          type: string
        emailConfirmado:
          type: boolean
        criadoEm:
          type: string
          format: date-time

    ReenvioConfirmacaoRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email

    Usuario:
      type: object
      description: >
        Dados públicos de um usuário (RF12/RF13) — nunca inclui senha_hash nem outros
        campos além dos listados; edição de perfil é escopo do Épico 4.
      required: [id, nome, email, perfil]
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        email:
          type: string
          format: email
        perfil:
          type: string
          enum: [ALUNO, MODERADOR]
          description: 'Perfil global do usuário (RF13).'
          examples:
            - ALUNO
        curso:
          type: [string, 'null']
          description: Curso informado no cadastro (Story 1.2) — usado na saudação da Home.

    ComunidadeRequest:
      type: object
      required: [nome]
      properties:
        nome:
          type: string
        descricao:
          type: string

    ComunidadeResponse:
      type: object
      required: [id, nome, tipo, criadoEm]
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        descricao:
          type: string
        tipo:
          type: string
          enum: [CURSO, ABERTA]
          description: 'Imutável após a criação (RF21.1).'
        souMembro:
          type: [boolean, 'null']
          description: >
            Se o usuário autenticado é membro desta comunidade (RF27.1) — presente só na
            consulta de uma comunidade específica, ausente nas listagens.
        criadoEm:
          type: string
          format: date-time

    ComunidadePageResponse:
      allOf:
        - $ref: '#/components/schemas/PageResponse'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/ComunidadeResponse'

    Erro:
      type: object
      description: >
        Envelope de erro padrão (AD-5) — toda resposta de erro de todo endpoint REST
        segue exatamente esta forma, com o status HTTP mapeado por cenário (401 sem
        autenticação, 403/404 conforme a existência do recurso deva ou não ficar
        oculta, 400/422 validação, 409 conflito, 500 erro não tratado).
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: 'Código estável da condição de erro (ex.: "NAO_AUTENTICADO").'
              examples:
                - NAO_AUTENTICADO
            message:
              type: string
              description: Mensagem legível para exibição/log.
              examples:
                - Autenticação necessária.
            details:
              type: string
              description: 'Detalhes adicionais opcionais (ex.: campo inválido).'

    PageResponse:
      type: object
      description: >
        Envelope de paginação padrão (AD-4) — todo endpoint de listagem referencia
        este componente via $ref, nunca inventa a própria forma de paginar. Como
        OpenAPI não tem genéricos, um endpoint concreto compõe o tipo do item com
        allOf, por exemplo:
          ComunidadePageResponse:
            allOf:
              - $ref: '#/components/schemas/PageResponse'
              - type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/Comunidade'
      required:
        - content
        - page
        - size
        - totalElements
        - totalPages
      properties:
        content:
          type: array
          description: Itens da página atual — cada endpoint concreto especifica o tipo real via allOf.
          items: {}
        page:
          type: integer
          format: int32
          minimum: 0
          description: Número da página atual (0-based).
        size:
          type: integer
          format: int32
          minimum: 1
          maximum: 100
          description: Quantidade máxima de itens por página.
        totalElements:
          type: integer
          format: int64
          minimum: 0
          description: Total de itens em todas as páginas.
        totalPages:
          type: integer
          format: int32
          minimum: 0
          description: Total de páginas disponíveis.

  responses:
    NaoAutenticado:
      description: Requisição sem autenticação válida (401).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Erro'
    SemPermissao:
      description: Autenticado, mas sem permissão para a ação (403).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Erro'
    NaoEncontrado:
      description: Recurso inexistente, ou cuja existência deve ficar oculta (404).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Erro'
    ErroValidacao:
      description: Corpo da requisição inválido (400/422).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Erro'
    Conflito:
      description: Conflito de estado do recurso (409).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Erro'
    ErroInterno:
      description: Erro não tratado (500).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Erro'

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Filtro de segurança JWT global (AD-2) — transporte exclusivamente via header
        Authorization: Bearer, nunca cookie. Claims fixos: sub (id do usuário) e
        roles (perfis globais, RF13).

security:
  - bearerAuth: []